medml-backend/requirements.txt
# HealthCare App/medml-backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Bcrypt
Flask-Cors
python-dotenv
pydantic
scikit-learn
pandas
joblib
xgboost
lightgbm
imbalanced-learn
fpdf # <-- ADDED for PDF generation